IMPLEMENT: use claude-haiku-4-5-20251001 for reconcile queries (closes #234)
Reconcile:query was sharing currentModel (claude-sonnet-4-6) with the chat
session. Haiku is sufficient for structured JSON extraction and significantly
faster. Extracted buildReconcileSystemPrompt, buildReconcileUserTurn, and
extractOpsArray into reconcile-query.ts with RECONCILE_MODEL = haiku constant.

Files changed:
- apps/desktop/src/main/ipc/reconcile-query.ts (new — pure helpers + model constant)
- apps/desktop/src/main/ipc/claude.ts (import from reconcile-query, use RECONCILE_MODEL)
- apps/desktop/tests/main/reconcile-query.test.ts (new — covers all pure helpers)
